refactor: 커스텀 애노테이션 관련 파일 위치 변경
