+++ src/test/java/com/example/spinlog/user/custom/securitycontext/WithMockCustomOAuth2User.java
@@ -1,17 +1,15 @@
-package com.example.spinlog.custom.securitycontext;
+package com.example.spinlog.user.custom.securitycontext;
 
 import org.springframework.security.test.context.support.WithSecurityContext;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
-import static com.example.spinlog.custom.securitycontext.OAuth2Provider.KAKAO;
-
 @Retention(RetentionPolicy.RUNTIME)
 @WithSecurityContext(factory = WithMockCustomOAuth2UserFactory.class)
 public @interface WithMockCustomOAuth2User {
 
-    OAuth2Provider provider() default KAKAO;
+    OAuth2Provider provider() default OAuth2Provider.KAKAO;
 
     String email() default "test@example.com";
